requirements.txt.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector==1.2.4
colorama==0.4.6
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.98.0
frozenlist==1.3.3
google-auth==2.21.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
orjson==3.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pyodbc==4.0.39
python-tds==1.12.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
yarl==1.9.2